Optimize kwargs filter for less calls to signature()
Before this, if you had large datasets with a lot of objects, a huge
part of the runtime went into inspect.signature() which was called for
every object times every encoder.

As the signatures of the encoders does not change, we can simplify this
and get the signatures only once. This reduced the runtime on a 1mb test
dataset from about 15s to 2s for encoding.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>